Reorganize stately code into separate modules
