+++ stately/.gitignore
@@ -1,0 +1,1 @@
+deps
