+++ stately/plugins/example.py
